% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhrBaseIntervalSubset.R
\name{rhrBaseIntervalSubset}
\alias{rhrBaseIntervalSubset}
\title{Interval Check}
\usage{
rhrBaseIntervalSubset(ts, int)
}
\arguments{
\item{ts}{Numeric vector, the time stamps in seconds.}

\item{int}{Numeric scalar, the minimum time in seconds that two observations need to be separated.}
}
\value{
A logcial vector of length \code{length(ts)}, indicating if a given observation is seperated from the last chosen one by int.
}
\description{
Checks if consecutive observations are separated by an interval or not. If \code{int} is used up, than the next observation is used.
}
\examples{
## sequence of 10 seconds
x <- seq(0, 3600, by = 60)
x[rhrBaseIntervalSubset(x, 600)]

## make the sequence irregular
x <- sort(sample(x, 40))
x[rhrBaseIntervalSubset(x, 600)]
}

