% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhrBaseMSD.R
\name{rhrBaseMSD}
\alias{rhrBaseMSD}
\title{Mean squared distance}
\usage{
rhrBaseMSD(x, y)
}
\arguments{
\item{x}{Numeric vector, the x cooridnates}

\item{y}{Numeric vector, the y coordinates}
}
\value{
Numeric matrix with the x and y coordinates of the new path.
}
\description{
Mean squared distance from the center of activity.
}

