% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhrBasePRW.R
\name{rhrBasePRW}
\alias{rhrBasePRW}
\title{Permutate walk}
\usage{
rhrBasePRW(x, y)
}
\arguments{
\item{x}{Numeric vector, the x coordinates.}

\item{y}{Numeric vector, the y coordinates.}
}
\value{
Numeric matrix with the x and y coordinates of the new path.
}
\description{
permuates a walk by reusing the original distances between steps and connecting them with angles drawn from a uniform distribution between 0 and 360.
}

